+++ src/libpgcode/src/Slic3r/Biz/libpgcode/ProcessorImpl.cpp
@@ -128,7 +128,7 @@ ProcessorResult ProcessorImpl::finalize()
         moves.resize(moves.size() + mod_moves.size());
         for (int i = int(moves.size()) - 1; i >= 0; --i) {
             moves[i] = moves[orig_idx];
-            if (! mod_moves.empty() && mod_moves.back() == orig_idx) {
+            if (! mod_moves.empty() && static_cast<int>(mod_moves.back()) == orig_idx) {
                 // What we just copied should be preceded by its copy (except for some fields).
                 moves[i-1] = moves[i];
                 MoveVertex& v = moves[i-1];
